File scoped namespaces, move tests into separate files, attempt different testing url for github actions
