+++ RealtimeTests/Models/Todo.cs
@@ -3,19 +3,18 @@ using System.Collections.Generic;
 using Postgrest.Attributes;
 using Postgrest.Models;
 
-namespace RealtimeTests.Models
+namespace RealtimeTests.Models;
+
+[Table("todos")]
+public class Todo : BaseModel
 {
-    [Table("todos")]
-    public class Todo : BaseModel
-    {
-        [PrimaryKey("id")] public int Id { get; set; }
+    [PrimaryKey("id")] public int Id { get; set; }
 
-        [Column("details")] public string? Details { get; set; }
+    [Column("details")] public string? Details { get; set; }
 
-        [Column("user_id")] public int UserId { get; set; }
+    [Column("user_id")] public int UserId { get; set; }
 
-        [Column("numbers")] public List<int>? Numbers { get; set; } = new List<int>();
+    [Column("numbers")] public List<int>? Numbers { get; set; } = new List<int>();
 
-        [Column("inserted_at")] public DateTime? InsertedAt { get; set; }
-    }
+    [Column("inserted_at")] public DateTime? InsertedAt { get; set; }
 }
