+++ RealtimeTests/Models/User.cs
@@ -1,15 +1,14 @@
 ﻿using Postgrest.Attributes;
 using Postgrest.Models;
 
-namespace RealtimeTests.Models
-{
-    [Table("users")]
-    public class User : BaseModel
-    {
-        [PrimaryKey("id", false)]
-        public string? Id { get; set; }
+namespace RealtimeTests.Models;
 
-        [Column("name")]
-        public string? Name { get; set; }
-    }
+[Table("users")]
+public class User : BaseModel
+{
+    [PrimaryKey("id", false)]
+    public string? Id { get; set; }
+
+    [Column("name")]
+    public string? Name { get; set; }
 }
